Replace fzf with CtrlP
fzf dirties the buffer list by creating a new buffer each time `:Files` is called. This quickly leads to large buffer counts (20+) that are difficult to navigate with `:bN`. Replacing this with CtrlP fixes this.

Also added a couple configuration changes to get CtrlP working about the same as fzf.